requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pillow==10.1.0
aiofiles==23.2.1
python-multipart==0.0.6
requests==2.31.0